BUG: One of our triangles were not possible to exist.
For a triangle to exist, the sum of two sides must be larger than the third side.

So, A + B must be higher than C and the process goes on for each side.

The bug was fixed, another test was created to validate the triangle and all
broken tests were fixed as well.